chore: fix response helper import
