+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -162,6 +162,13 @@
 - **Completion Date**: 2025-08-24
 - **Summary**: Brought down from 892 to 148 lines by modularizing handshake, routing, authentication, and logging utilities. Protocol file now orchestrates these modules.
 
+### MAJOR-008: Unified Financial API Refactor ✅
+- **File**: `src/services/financial/unified_financial_api.py`
+- **Status**: Completed
+- **Assigned To**: Agent
+- **Completion Date**: 2025-08-24
+- **Summary**: Split monolithic unified financial API into router, processor, responses, and config modules. Orchestrator now coordinates these components.
+
 ### MODERATE-001: AI Development Workflow Refactor ✅
 - **File**: `src/ai_ml/dev_workflow.py`
 - **Status**: Completed
